[Refactor] Fix static analysis errors in DatabaseQueries
