+++ site/app/libraries/database/DatabaseQueries.php
@@ -9,16 +9,14 @@ use app\libraries\Core;
 use app\libraries\DateUtils;
 use app\libraries\ForumUtils;
 use app\libraries\GradeableType;
-use app\models\Gradeable;
 use app\models\gradeable\Component;
+use app\models\gradeable\Gradeable;
 use app\models\gradeable\GradedComponent;
 use app\models\gradeable\GradedGradeable;
 use app\models\gradeable\Mark;
 use app\models\gradeable\RegradeRequest;
 use app\models\gradeable\Submitter;
 use app\models\gradeable\TaGradedGradeable;
-use app\models\GradeableComponent;
-use app\models\GradeableComponentMark;
 use app\models\User;
 use app\models\Notification;
 use app\models\SimpleLateUser;
@@ -1697,102 +1695,6 @@ VALUES(?, ?, ?, ?, 0, 0, 0, 0, ?)", array($g_id, $user_id, $team_id, $version, $
         }, $this->course_db->rows());
     }
 
-    public function insertGradeableComponentMarkData($gd_id, $gc_id, $gcd_grader_id, GradeableComponentMark $mark) {
-        $params = array($gc_id, $gd_id, $gcd_grader_id, $mark->getId());
-        $this->course_db->query("
-INSERT INTO gradeable_component_mark_data (gc_id, gd_id, gcd_grader_id, gcm_id)
-VALUES (?, ?, ?, ?)", $params);
-    }
-
-    public function createNewGradeableComponent(GradeableComponent $component, $gradeable_id) {
-        $params = array($gradeable_id, $component->getTitle(), $component->getTaComment(),
-                        $component->getStudentComment(), $component->getLowerClamp(), $component->getDefault(),
-                        $component->getMaxValue(), $component->getUpperClamp(),
-                        $this->course_db->convertBoolean($component->getIsText()), $component->getOrder(),
-                        $this->course_db->convertBoolean($component->getIsPeer()), $component->getPage());
-        $this->course_db->query("
-INSERT INTO gradeable_component(g_id, gc_title, gc_ta_comment, gc_student_comment, gc_lower_clamp, gc_default, gc_max_value, gc_upper_clamp,
-gc_is_text, gc_order, gc_is_peer, gc_page)
-VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)", $params);
-    }
-
-    public function updateGradeableComponent(GradeableComponent $component) {
-        $params = array($component->getTitle(), $component->getTaComment(), $component->getStudentComment(),
-                        $component->getLowerClamp(), $component->getDefault(), $component->getMaxValue(),
-                        $component->getUpperClamp(), $this->course_db->convertBoolean($component->getIsText()),
-                        $component->getOrder(), $this->course_db->convertBoolean($component->getIsPeer()),
-                        $component->getPage(), $component->getId());
-        $this->course_db->query("
-UPDATE gradeable_component SET gc_title=?, gc_ta_comment=?, gc_student_comment=?, gc_lower_clamp=?, gc_default=?, gc_max_value=?, gc_upper_clamp=?, gc_is_text=?, gc_order=?, gc_is_peer=?, gc_page=? WHERE gc_id=?", $params);
-    }
-
-    public function deleteGradeableComponent(GradeableComponent $component) {
-        $this->course_db->query("DELETE FROM gradeable_component_data WHERE gc_id=?",array($component->getId()));
-        $this->course_db->query("DELETE FROM gradeable_component WHERE gc_id=?", array($component->getId()));
-    }
-
-    public function createGradeableComponentMark(GradeableComponentMark $mark) {
-        $bool_value = $this->course_db->convertBoolean($mark->getPublish());
-        $params = array($mark->getGcId(), $mark->getPoints(), $mark->getNoteNoDecode(), $mark->getOrder(), $bool_value);
-
-        $this->course_db->query("
-INSERT INTO gradeable_component_mark (gc_id, gcm_points, gcm_note, gcm_order, gcm_publish)
-VALUES (?, ?, ?, ?, ?)", $params);
-        return $this->course_db->getLastInsertId();
-    }
-
-    public function updateGradeableComponentMark(GradeableComponentMark $mark) {
-        $bool_value = $this->course_db->convertBoolean($mark->getPublish());
-        $params = array($mark->getGcId(), $mark->getPoints(), $mark->getNoteNoDecode(), $mark->getOrder(), $bool_value, $mark->getId());
-        $this->course_db->query("
-UPDATE gradeable_component_mark SET gc_id=?, gcm_points=?, gcm_note=?, gcm_order=?, gcm_publish=?
-WHERE gcm_id=?", $params);
-    }
-
-    public function deleteGradeableComponentMark(GradeableComponentMark $mark) {
-        $this->course_db->query("DELETE FROM gradeable_component_mark_data WHERE gcm_id=?",array($mark->getId()));
-        $this->course_db->query("DELETE FROM gradeable_component_mark WHERE gcm_id=?", array($mark->getId()));
-    }
-
-    public function getGreatestGradeableComponentMarkOrder(GradeableComponent $component) {
-    	$this->course_db->query("SELECT MAX(gcm_order) as max FROM gradeable_component_mark WHERE gc_id=? ", array($component->getId()));
-    	$row = $this->course_db->row();
-        return $row['max'];
-
-    }
-
-    /**
-     * This updates the viewed date on a gradeable object (assuming that it has a set
-     * $user object associated with it).
-     *
-     * @param \app\models\Gradeable $gradeable
-     */
-    public function updateUserViewedDate(Gradeable $gradeable) {
-        if ($gradeable->getGdId() !== null && $gradeable->getUser() !== null) {
-            $this->course_db->query("UPDATE gradeable_data SET gd_user_viewed_date = NOW() WHERE gd_id=? and gd_user_id=?",
-                array($gradeable->getGdId(), $gradeable->getUser()->getId()));
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * This updates the viewed date on a gradeable object (assuming that it has a set
-     * $user object associated with it).
-     *
-     * @param \app\models\Gradeable $gradeable
-     */
-    public function resetUserViewedDate(Gradeable $gradeable) {
-        if ($gradeable->getGdId() !== null && $gradeable->getUser() !== null) {
-            $this->course_db->query("UPDATE gradeable_data SET gd_user_viewed_date = NULL WHERE gd_id=? and gd_user_id=?",
-                array($gradeable->getGdId(), $gradeable->getUser()->getId()));
-            return true;
-        } else {
-            return false;
-        }
-    }
-
     /**
      * Finds the viewed time for a specific user on a team.
      * Assumes team_ids are unique (cannot be used for 2 different gradeables)
@@ -2122,9 +2024,9 @@ WHERE gcm_id=?", $params);
      * Return array of counts of teams/users without team/graded components
      * corresponding to each registration/rotating section
      * @param string $g_id
-     * @param rray(int) $sections
+     * @param int[] $sections
      * @param string $section_key
-     * @return array(int) $return
+     * @return int[] $return
      */
     public function getTotalTeamCountByGradingSections($g_id, $sections, $section_key) {
         $return = array();
@@ -4042,7 +3944,7 @@ AND gc_id IN (
      * @param Submitter $submitter
      * @return bool
      */
-    public function getHasSubmission(gradeable\Gradeable $gradeable, Submitter $submitter) {
+    public function getHasSubmission(Gradeable $gradeable, Submitter $submitter) {
         $this->course_db->query('SELECT EXISTS (SELECT g_id FROM electronic_gradeable_data WHERE g_id=? AND (user_id=? OR team_id=?))',
             [$gradeable->getId(), $submitter->getId(), $submitter->getId()]);
         return $this->course_db->row()['exists'] ?? false;
@@ -4055,7 +3957,7 @@ AND gc_id IN (
      * @param string[] $submitter_ids
      * @return bool[] Map of id=>version
      */
-    public function getActiveVersions(gradeable\Gradeable $gradeable, array $submitter_ids) {
+    public function getActiveVersions(Gradeable $gradeable, array $submitter_ids) {
         throw new NotImplementedException();
     }
 
@@ -4093,7 +3995,7 @@ AND gc_id IN (
      * @param Gradeable\Gradeable $gradeable
      * @return array
      */
-    public function getUsersNotFullyGraded(Gradeable\Gradeable $gradeable, $component_id = "-1") {
+    public function getUsersNotFullyGraded(Gradeable $gradeable, $component_id = "-1") {
 
         // Get variables needed for query
         $component_count = count($gradeable->getComponents());
